# 04_前端工程/mineworks-web/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MineWorks API",
    "description": "矿业智工平台工程计算API。计算核心、权限、数据库和HTTP层严格分离。",
    "version": "1.1.0"
  },
  "paths": {
...
    "/api/v1/billing/entitlements": {
      "get": {
        "tags": [
          "plans-entitlements-and-billing"
        ],
        "summary": "Entitlements",
        "operationId": "entitlements_api_v1_billing_entitlements_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitlementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/orders": {
      "get": {
        "tags": [
          "plans-entitlements-and-billing"
        ],
        "summary": "Orders",
        "operationId": "orders_api_v1_billing_orders_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BillingOrderResponse"
                  },
                  "title": "Response Orders Api V1 Billing Orders Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/subscriptions": {
      "get": {
        "tags": [
          "plans-entitlements-and-billing"
        ],
        "summary": "Subscriptions",
        "operationId": "subscriptions_api_v1_billing_subscriptions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubscriptionResponse"
                  },
                  "title": "Response Subscriptions Api V1 Billing Subscriptions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/checkout-intents": {
      "post": {
        "tags": [
          "plans-entitlements-and-billing"
        ],
        "summary": "Create Checkout Intent",
        "operationId": "create_checkout_intent_api_v1_billing_checkout_intents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Team-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutIntentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingOrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/billing/orders/{order_id}/simulate-paid": {
      "post": {
        "tags": [
          "plans-entitlements-and-billing"
        ],
        "summary": "Simulate Paid",
        "operationId": "simulate_paid_api_v1_billing_orders__order_id__simulate_paid_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "X-Team-Id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Team-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingOrderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingOrderResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "subject_type": {
            "type": "string",
            "enum": [
              "user",
              "team"
            ],
            "title": "Subject Type"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "target_plan": {
            "type": "string",
            "enum": [
              "free",
              "professional",
              "team",
              "enterprise"
            ],
            "title": "Target Plan"
          },
          "billing_cycle": {
            "type": "string",
            "enum": [
              "monthly",
              "annual"
            ],
            "title": "Billing Cycle"
          },
          "amount_fen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Fen"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "paid",
              "cancelled",
              "failed"
            ],
            "title": "Status"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "provider_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Reference"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "paid_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Paid At"
          },
          "payment_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Action"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "team_id",
          "subject_type",
          "subject_id",
          "target_plan",
          "billing_cycle",
          "amount_fen",
          "currency",
          "status",
          "provider",
          "provider_reference",
          "created_at",
          "updated_at",
          "paid_at"
        ],
        "title": "BillingOrderResponse"
      },
...
      "CheckoutIntentCreate": {
        "properties": {
          "subject_type": {
            "type": "string",
            "enum": [
              "user",
              "team"
            ],
            "title": "Subject Type"
          },
          "target_plan": {
            "type": "string",
            "enum": [
              "free",
              "professional",
              "team",
              "enterprise"
            ],
            "title": "Target Plan"
          },
          "billing_cycle": {
            "type": "string",
            "enum": [
              "monthly",
              "annual"
            ],
            "title": "Billing Cycle",
            "default": "monthly"
          }
        },
        "type": "object",
        "required": [
          "subject_type",
          "target_plan"
        ],
        "title": "CheckoutIntentCreate"
      },
...
      "EntitlementResponse": {
        "properties": {
          "user_plan": {
            "type": "string",
            "enum": [
              "free",
              "professional",
              "team",
              "enterprise"
            ],
            "title": "User Plan"
          },
          "team_plan": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "free",
                  "professional",
                  "team",
                  "enterprise"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Plan"
          },
          "effective_plan": {
            "type": "string",
            "enum": [
              "free",
              "professional",
              "team",
              "enterprise"
            ],
            "title": "Effective Plan"
          },
          "plan_source": {
            "type": "string",
            "enum": [
              "user",
              "team",
              "free"
            ],
            "title": "Plan Source"
          },
          "active_team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Team Id"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          },
          "quotas": {
            "items": {
              "$ref": "#/components/schemas/UsageQuotaResponse"
            },
            "type": "array",
            "title": "Quotas"
          },
          "billing_mode": {
            "type": "string",
            "const": "local-development",
            "title": "Billing Mode",
            "default": "local-development"
          },
          "commercial_ready": {
            "type": "boolean",
            "title": "Commercial Ready",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "user_plan",
          "team_plan",
          "effective_plan",
          "plan_source",
          "active_team_id",
          "features",
          "quotas"
        ],
        "title": "EntitlementResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SubscriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "subject_type": {
            "type": "string",
            "enum": [
              "user",
              "team"
            ],
            "title": "Subject Type"
          },
          "subject_id": {
            "type": "string",
            "title": "Subject Id"
          },
          "plan": {
            "type": "string",
            "enum": [
              "free",
              "professional",
              "team",
              "enterprise"
            ],
            "title": "Plan"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "active",
              "cancelled",
              "expired"
            ],
            "title": "Status"
          },
          "billing_cycle": {
            "type": "string",
            "enum": [
              "monthly",
              "annual"
            ],
            "title": "Billing Cycle"
          },
          "starts_at": {
            "type": "string",
            "title": "Starts At"
          },
          "ends_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ends At"
          },
          "auto_renew": {
            "type": "boolean",
            "title": "Auto Renew"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "external_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Reference"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "subject_type",
          "subject_id",
          "plan",
          "status",
          "billing_cycle",
          "starts_at",
          "ends_at",
          "auto_renew",
          "provider",
          "external_reference",
          "created_at",
          "updated_at"
        ],
        "title": "SubscriptionResponse"
      },
...
      "UsageQuotaResponse": {
        "properties": {
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "period": {
            "type": "string",
            "enum": [
              "monthly",
              "total"
            ],
            "title": "Period"
          },
          "used": {
            "type": "integer",
            "title": "Used"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "remaining": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remaining"
          },
          "percentage": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Percentage"
          },
          "period_start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Start"
          },
          "period_end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period End"
          }
        },
        "type": "object",
        "required": [
          "metric",
          "label",
          "period",
          "used",
          "limit",
          "remaining",
          "percentage"
        ],
        "title": "UsageQuotaResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}